requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
coverage==7.4.1
Django==3.2.17
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==22.3
django-chartjs==2.3.0
django-ckeditor==6.5.1
django-ipware==3.0.7
django-js-asset==2.0.0
django-paypal==2.0
django-stdimage==6.0.1
gunicorn==20.1.0
model-bakery==1.17.0
model-mommy==2.0.0
mysqlclient==2.2.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pydantic==1.10.7
PyMySQL==1.0.3
pytz==2023.3
soupsieve==2.4
sqlparse==0.4.3
style==1.1.0
typing_extensions==4.5.0
update==0.0.1
whitenoise==6.4.0